+++ package.json
@@ -25,14 +25,15 @@
 		"vuepress": "^1.3.1",
 		"vuepress-plugin-clean-urls": "^1.1.1",
 		"vuepress-plugin-container": "^2.1.2",
-		"vuepress-types": "^0.9.0"
+		"vuepress-types": "^0.9.2"
 	},
 	"dependencies": {
+		"@lottiefiles/lottie-player": "^0.5.1",
 		"axios": "^0.19.2",
 		"eslint-plugin-jsx-a11y": "^6.2.3",
 		"eslint-plugin-react": "^7.19.0",
-		"eslint-plugin-react-hooks": "^2.5.0",
+		"eslint-plugin-react-hooks": "^2.5.1",
 		"prettier-config-vuepress": "^1.0.0",
-		"vuepress-theme-yuu": "^2.2.1"
+		"vuepress-theme-yuu": "^2.3.0"
 	}
 }
