+++ src/.vuepress/styles/index.styl
@@ -11,11 +11,38 @@ font-feature-settings()
 	-ms-font-feature-settings arguments
 	font-feature-settings arguments
 
+// Yuu theme changes
+.yuu-theme-dark
+	a.settings-button
+		color $darkTextColor
+		&:hover
+			color $darkAltTextColor
+	.nav-links
+		.nav-item
+			a.nav-link
+				&.router-link-active
+					color $darkTextColor
+				&:hover
+					color $darkAltTextColor
+	.dropdown-wrapper
+		.nav-dropdown
+			li
+				&.dropdown-item a
+					color $darkAltTextColor
+					&.router-link-active,
+					&:not(.router-link-active)
+						&:hover
+							color $darkTextColor
+				h4
+					border-top 1px solid $darkSecondaryBg
+				&:first-child h4
+					border-top 0
+
 // Custom block containers
 .custom-block
-	&.tip
-	&.warning
-	&.error
+	&.tip,
+	&.warning,
+	&.error,
 	&.danger
 		border-radius $containerBorderRadius
 	&.aside
@@ -26,7 +53,7 @@ font-feature-settings()
 		font-size 0.9rem
 		text-align left
 		p
-			margin-top: 0
+			margin-top 0
 		a
 			color $accentColorSecondary
 
@@ -69,7 +96,7 @@ font-feature-settings()
 .headerLogo
 	vertical-align middle
 	height 3rem
-	margin-bottom .625rem
+	margin-bottom 0.625rem
 
 // Center align class
 .centered
