+++ src/.vuepress/styles/palette.styl
@@ -15,7 +15,9 @@ $navbarHeight = 3.6rem
 $sidebarWidth = 20rem
 $contentWidth = 46.25rem
 $homePageWidth = 70rem
-$heroImageHeight = 8rem
+$heroImageHeight = 11rem
+$heroTitleSize = 3rem
+$heroDescSize = 1.6rem
 $containerBorderRadius = 0.4rem
 $buttonBorderRadius = 0.25rem
 // Responsive breakpoints
