--- src/.vuepress/theme/components/Home.vue
@@ -1,220 +1,0 @@
-<template>
-  <main
-    class="home"
-    aria-labelledby="main-title"
-  >
-    <header class="hero">
-      <img
-        v-if="data.heroImage"
-        :src="$withBase(data.heroImage)"
-        :alt="data.heroAlt || 'hero'"
-      >
-
-      <h1
-        v-if="data.heroText !== null"
-        id="main-title"
-      >
-        {{ data.heroText || $title || 'Paperback' }}
-      </h1>
-
-      <p
-        v-if="data.tagline !== null"
-        class="description"
-      >
-        {{ data.tagline || $description || 'An ad-free manga reader for iOS.' }}
-      </p>
-
-      <p
-        v-if="data.actionText && data.actionLink && data.actionDownload && browserDownloadUrl"
-        class="action"
-      >
-        <NavLink
-          class="action-button action-button-download"
-          :item="actionDownload"
-        />
-        <NavLink
-          class="action-button action-button-getstarted"
-          :item="actionLink"
-        />
-      </p>
-    </header>
-
-    <div
-      v-if="data.features && data.features.length"
-      class="features"
-    >
-      <div
-        v-for="(feature, index) in data.features"
-        :key="index"
-        class="feature"
-      >
-        <h2>{{ feature.title }}</h2>
-        <p>{{ feature.details }}</p>
-      </div>
-    </div>
-
-    <Content class="theme-default-content custom" />
-
-    <div
-      v-if="data.footer"
-      class="footer"
-    >
-      {{ data.footer }}
-    </div>
-  </main>
-</template>
-
-<script>
-import NavLink from '@theme/components/NavLink.vue'
-import axios from "axios";
-
-const RELEASE_URL =
-  "https://api.github.com/repos/Paperback-iOS/app/releases/latest";
-
-const LATEST_RELEASE =
-  "https://github.com/Paperback-iOS/app/releases/latest";
-
-export default {
-  name: 'Home',
-
-  components: { NavLink },
-
-	data() {
-		return {
-			tagName: "",
-			browserDownloadUrl: "",
-		};
-  },
-
-  computed: {
-    data () {
-      return this.$page.frontmatter
-    },
-
-    actionDownload () {
-      return {
-        link: this.$data.browserDownloadUrl || LATEST_RELEASE,
-        text: this.data.actionDownload
-      }
-    },
-
-    actionLink () {
-      return {
-        link: this.data.actionLink,
-        text: this.data.actionText
-      }
-    }
-  },
-
-	async mounted() {
-		const { data } = await axios.get(RELEASE_URL);
-		// Maybe eventually some release has more than the ipa in assets.
-		const ipaAsset = data.assets.find((a) => a.name.includes(".ipa"));
-		// Set the values.
-		this.$data.tagName = data.tag_name;
-		this.$data.browserDownloadUrl = ipaAsset.browser_download_url;
-  },
-
-}
-</script>
-
-<style lang="stylus">
-.home
-  padding $navbarHeight 2rem 0
-  max-width $homePageWidth
-  margin 0px auto
-  display block
-  .hero
-    text-align center
-    img
-      max-width: 100%
-      max-height $heroImageHeight
-      display block
-      margin 3rem auto 1.5rem
-    h1
-      font-size 3rem
-    h1, .description, .action
-      margin 1.8rem auto
-    .description
-      max-width 35rem
-      font-size 1.6rem
-      line-height 1.3
-      color lighten($textColor, 40%)
-    .action-button
-      margin 0.25rem
-      min-width 11rem
-      display inline-block
-      font-size 1.2rem
-      color #fff
-      padding 0.8rem 1.6rem
-      border-radius 4px
-      transition background-color .1s ease
-      box-sizing border-box
-      .icon.outbound
-        display none
-      &.action-button-download
-        background-color $accentColor
-        border-bottom 1px solid darken($accentColor, 10%)
-        &:hover
-          background-color lighten($accentColor, 10%)
-      &.action-button-getstarted
-        background-color darken($accentColor, 30%)
-        border-bottom 1px solid darken($accentColor, 40%)
-        &:hover
-          background-color darken($accentColor, 20%)
-  .features
-    border-top 1px solid $borderColor
-    padding 1.2rem 0
-    margin-top 2.5rem
-    display flex
-    flex-wrap wrap
-    align-items flex-start
-    align-content stretch
-    justify-content space-between
-  .feature
-    flex-grow 1
-    flex-basis 30%
-    max-width 30%
-    h2
-      font-size 1.4rem
-      font-weight 500
-      border-bottom none
-      padding-bottom 0
-      color lighten($textColor, 10%)
-    p
-      color lighten($textColor, 25%)
-  .footer
-    padding 2.5rem
-    border-top 1px solid $borderColor
-    text-align center
-    color lighten($textColor, 25%)
-
-@media (max-width: $MQMobile)
-  .home
-    .features
-      flex-direction column
-    .feature
-      max-width 100%
-      padding 0 2.5rem
-
-@media (max-width: $MQMobileNarrow)
-  .home
-    padding-left 1.5rem
-    padding-right 1.5rem
-    .hero
-      img
-        max-height 210px
-        margin 2rem auto 1.2rem
-      h1
-        font-size 2rem
-      h1, .description, .action
-        margin 1.2rem auto
-      .description
-        font-size 1.2rem
-      .action-button
-        font-size 1rem
-        padding 0.6rem 1.2rem
-    .feature
-      h2
-        font-size 1.25rem
-</style>
