--- src/.vuepress/theme/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-	extend: "@vuepress/theme-default",
-};
