--- src/.vuepress/theme/styles/themes/blue.styl
@@ -1,143 +1,0 @@
-@import '../variables.styl';
-
-.yuu-theme-blue {
-	a,
-	p a code,
-	.theme-default-content a code {
-		color: $blueAccentColor;
-	}
-
-	// @parent-theme/components/AlgoliaSearchBox.vue
-	.algolia-search-wrapper .algolia-autocomplete {
-		.algolia-docsearch-suggestion {
-			.algolia-docsearch-suggestion--category-header {
-				background-color: darken($blueAccentColor, 10%);
-			}
-
-			.algolia-docsearch-suggestion--subcategory-column {
-				.algolia-docsearch-suggestion--highlight {
-					color: $blueAccentColor;
-				}
-			}
-		}
-
-		.algolia-docsearch-suggestion--text .algolia-docsearch-suggestion--highlight {
-			box-shadow: inset 0 -2px 0 0 rgba($blueAccentColor, 0.8);
-		}
-
-		.algolia-docsearch-suggestion--highlight {
-			color: $blueAccentColor;
-		}
-	}
-
-	// @parent-theme/components/Home.vue
-	.hero .action-button {
-		&.action-button-download {
-			background-color: $blueAccentColor;
-			border-bottom-color: darken($blueAccentColor, 10%);
-			&:hover {
-				background-color: lighten($blueAccentColor, 10%);
-			}
-		}
-		&.action-button-getstarted {
-			background-color: darken($blueAccentColor, 30%);
-			border-bottom-color: darken($blueAccentColor, 40%);
-			&:hover {
-				background-color: darken($blueAccentColor, 20%);
-			}
-		}
-	}
-
-	// @parent-theme/components/NavLinks.vue
-	.nav-links {
-		a,
-		.dropdown-wrapper .dropdown-title,
-		.dropdown-wrapper .nav-dropdown .dropdown-item a {
-			color: inherit;
-
-			&:hover,
-			&.router-link-active {
-				color: $blueAccentColor;
-			}
-		}
-
-		.dropdown-wrapper .nav-dropdown .dropdown-item a.router-link-active::after {
-			border-left-color: $blueAccentColor;
-		}
-	}
-
-	@media (min-width: $MQMobile) {
-		.nav-item > a:not(.external) {
-			&:hover,
-			&.router-link-active {
-				border-bottom-color: $blueAccentColor;
-			}
-		}
-	}
-
-	// @parent-theme/components/PageEdit.vue
-	.page-edit {
-		.edit-link a {
-			color: $blueAccentColor;
-		}
-
-		.last-updated .prefix {
-			color: $blueAccentColor;
-		}
-	}
-
-	// @vuepress/plugin-search/SearchBox.vue
-	.search-box {
-		input:focus {
-			border-color: alpha($blueAccentColor, 0.75);
-		}
-
-		.suggestion {
-			&.focused a {
-				color: $blueAccentColor;
-			}
-		}
-	}
-
-	// @parent-theme/components/SidebarLink.vue
-	&.yuu-theme-dark {
-		a.sidebar-link {
-			color: $darkAltTextColor;
-		}
-	}
-
-	// @parent-theme/components/SidebarLink.vue
-	a.sidebar-link {
-		color: $textColor;
-
-		&:hover {
-			color: $blueAccentColor;
-		}
-
-		&.active {
-			color: $blueAccentColor;
-			border-left-color: $blueAccentColor;
-		}
-	}
-
-	// @vuepress/plugin-pwa/lib/SWUpdatePopup.vue
-	.sw-update-popup {
-		border-color: $blueAccentColor;
-	}
-
-	// @vuepress/plugin-nprogress/nprogress.styl
-	#nprogress {
-		.bar {
-			background-color: $blueAccentColor;
-		}
-
-		.peg {
-			box-shadow: 0 0 10px $blueAccentColor, 0 0 5px $blueAccentColor;
-		}
-
-		.spinner-icon {
-			border-top-color: $blueAccentColor;
-			border-left-color: $blueAccentColor;
-		}
-	}
-}
