--- src/.vuepress/theme/styles/themes/purple.styl
@@ -1,143 +1,0 @@
-@import '../variables.styl';
-
-.yuu-theme-purple {
-	a,
-	p a code,
-	.theme-default-content a code {
-		color: $purpleAccentColor;
-	}
-
-	// @parent-theme/components/AlgoliaSearchBox.vue
-	.algolia-search-wrapper .algolia-autocomplete {
-		.algolia-docsearch-suggestion {
-			.algolia-docsearch-suggestion--category-header {
-				background-color: darken($purpleAccentColor, 10%);
-			}
-
-			.algolia-docsearch-suggestion--subcategory-column {
-				.algolia-docsearch-suggestion--highlight {
-					color: $purpleAccentColor;
-				}
-			}
-		}
-
-		.algolia-docsearch-suggestion--text .algolia-docsearch-suggestion--highlight {
-			box-shadow: inset 0 -2px 0 0 rgba($purpleAccentColor, 0.8);
-		}
-
-		.algolia-docsearch-suggestion--highlight {
-			color: $purpleAccentColor;
-		}
-	}
-
-	// @parent-theme/components/Home.vue
-	.hero .action-button {
-		&.action-button-download {
-			background-color: $purpleAccentColor;
-			border-bottom-color: darken($purpleAccentColor, 10%);
-			&:hover {
-				background-color: lighten($purpleAccentColor, 10%);
-			}
-		}
-		&.action-button-getstarted {
-			background-color: darken($purpleAccentColor, 30%);
-			border-bottom-color: darken($purpleAccentColor, 40%);
-			&:hover {
-				background-color: darken($purpleAccentColor, 20%);
-			}
-		}
-	}
-
-	// @parent-theme/components/NavLinks.vue
-	.nav-links {
-		a,
-		.dropdown-wrapper .dropdown-title,
-		.dropdown-wrapper .nav-dropdown .dropdown-item a {
-			color: inherit;
-
-			&:hover,
-			&.router-link-active {
-				color: $purpleAccentColor;
-			}
-		}
-
-		.dropdown-wrapper .nav-dropdown .dropdown-item a.router-link-active::after {
-			border-left-color: $purpleAccentColor;
-		}
-	}
-
-	@media (min-width: $MQMobile) {
-		.nav-item > a:not(.external) {
-			&:hover,
-			&.router-link-active {
-				border-bottom-color: $purpleAccentColor;
-			}
-		}
-	}
-
-	// @parent-theme/components/PageEdit.vue
-	.page-edit {
-		.edit-link a {
-			color: $purpleAccentColor;
-		}
-
-		.last-updated .prefix {
-			color: $purpleAccentColor;
-		}
-	}
-
-	// @vuepress/plugin-search/SearchBox.vue
-	.search-box {
-		input:focus {
-			border-color: alpha($purpleAccentColor, 0.75);
-		}
-
-		.suggestion {
-			&.focused a {
-				color: $purpleAccentColor;
-			}
-		}
-	}
-
-	// @parent-theme/components/SidebarLink.vue
-	&.yuu-theme-dark {
-		a.sidebar-link {
-			color: $darkAltTextColor;
-		}
-	}
-
-	// @parent-theme/components/SidebarLink.vue
-	a.sidebar-link {
-		color: $textColor;
-
-		&:hover {
-			color: $purpleAccentColor;
-		}
-
-		&.active {
-			color: $purpleAccentColor;
-			border-left-color: $purpleAccentColor;
-		}
-	}
-
-	// @vuepress/plugin-pwa/lib/SWUpdatePopup.vue
-	.sw-update-popup {
-		border-color: $purpleAccentColor;
-	}
-
-	// @vuepress/plugin-nprogress/nprogress.styl
-	#nprogress {
-		.bar {
-			background-color: $purpleAccentColor;
-		}
-
-		.peg {
-			box-shadow: 0 0 10px $purpleAccentColor, 0 0 5px $purpleAccentColor;
-		}
-
-		.spinner-icon {
-			border-top-color: $purpleAccentColor;
-			border-left-color: $purpleAccentColor;
-		}
-	}
-}
