--- src/.vuepress/theme/styles/variables.styl
@@ -1,25 +1,0 @@
-// overwrites
-$contentWidth = 920px;
-
-// dark theme colors
-$darkPrimaryBg = #1a1a1a;
-$darkSecondaryBg = #2a2a2a;
-$darkTextColor = #f3f3f3;
-$darkAltTextColor = #b9b9b9;
-$darkBorderColor = alpha(#fff, 0.1);
-
-$purple = #dd6efd;
-
-// red theme colors
-$redAccentColor = #de3636;
-
-// blue theme colors
-$blueAccentColor = #2196f3;
-
-// purple theme colors
-$purpleAccentColor = #a846eb;
-
-// content block colors
-$tipGreen = #42b983;
-$warningYellow = #ffe564;
-$dangerRed = #c00;
